iPAINT-paintvendingmachine-main/iPAINT-paintvendingmachine-main/requirements.txt
firebase-admin
customtkinter
Pillow
opencv-python
pyserial